finish removing file timeline map
