+++ src/server/timelines/game-data.ts
@@ -12,11 +12,17 @@ import { PRE_CPIP_STATIC_FILES } from "@server/game-data/precpip-static";
 import { RoomName } from "@server/game-data/rooms";
 import { ORIGINAL_STAMPBOOK, Stampbook } from "@server/game-data/stamps";
 import { ItemTable } from "@server/game-logic/items";
+import { getNewspaperName } from "@server/routes/news.txt";
 import { isGreater, Version } from "@server/routes/versions";
 import { SettingsManager } from "@server/settings";
 import { CatalogItems, CPUpdateE, CrumbIndicator, GameUpdate, HuntCrumbs } from "@server/updates";
 import path from "path";
 import { SCAVENGER_ICON_PATH, TICKET_INFO_PATH } from "./crumbs";
+import { NEWSPAPER_TIMELINE } from "./newspapers";
+
+export function getMinifiedDate(date: Version): string {
+  return date.replaceAll('-', '');
+}
 
 type GameState = {
   /** map of route -> path of file in media folder, meant for static files */
@@ -69,7 +75,7 @@ export class GameData {
   }
 
   private addRoute(route: string, file: FileRef) {
-    this.state.files.set(route, getMediaFilePath(file));
+    this.state.files.set(route.replaceAll('\\', '/'), getMediaFilePath(file));
   }
 
   private addCatalog(input: FileRef | CatalogItems, paths: string[]) {
@@ -94,6 +100,79 @@ export class GameData {
   private addRouteMap(routeMap: RouteRefMap): void {
     iterateEntries(routeMap, (route, file) => {
       this.addRoute(route, file);
+    });
+  }
+
+  private addNewspapers(): void {
+    const configXmlPath = getMediaFilePath('tool:news_config.xml');
+    NEWSPAPER_TIMELINE.forEach((update, i) => {
+      if (typeof update.info === 'string' || 'file' in update.info) {
+        const file = typeof update.info === 'string' ? update.info : update.info.file;
+        const issue = i + 1;
+
+        // pre-cpip, before rewrite
+        this.addRoute(`artwork/news/news${issue}.swf`, file);
+        // pre-cpip, post rewrite
+        const route2007 = getNewspaperName(update.date).replace('|', '/') + '.swf';
+        this.addRoute(path.join('artwork/news', route2007), file);
+
+        // 2006 boiler room (likely inaccurate, this artwork/archives was probably not a newspaper but a bundle of papers)
+        this.addRoute(path.join('artwork/archives', `news${issue + 1}.swf`), file);
+
+        // post-cpip
+        const date = getMinifiedDate(update.date);
+        this.addRoute(`play/v2/content/local/en/news/${date}/${date}.swf`, file);
+      } else {
+        const baseNewsPath = 'play/v2/content/local/en/news/';
+        const oldNewsPath = `${baseNewsPath}${getMinifiedDate(update.date)}`;
+        const newNewsPath = `${baseNewsPath}papers/${getMinifiedDate(update.date)}`;
+        this.addRoute(path.join(oldNewsPath, 'config.xml'), configXmlPath);
+        this.addRoute(path.join(newNewsPath, 'config.xml'), configXmlPath);
+        const newspaperComponenets: Array<[string, string]> = [
+          ['front/header.swf', update.info.headerFront ?? 'archives:News285HeaderFront.swf'],
+          ['front/featureStory.swf', update.info.featureStory],
+          ['front/supportStory.swf', update.info.supportStory],
+          ['front/upcomingEvents.swf', update.info.upcomingEvents],
+          ['front/newsFlash.swf', update.info.newsFlash],
+          ['front/askAuntArctic.swf', update.info.askFront],
+          ['front/dividers.swf', update.info.dividersFront ?? 'approximation:dividers_blank.swf'],
+          ['front/navigation.swf', update.info.navigationFront ?? 'archives:News268NavigationFront.swf'],
+          ['back/header.swf', update.info.headerBack ?? 'archives:News285HeaderBack.swf'],
+          ['back/askAuntArctic.swf', update.info.askBack],
+          ['back/secrets.swf', update.info.secrets ?? 'archives:News285Secrets.swf'],
+          ['back/submitYourContent.swf', update.info.submit ?? 'archives:News268SubmitYourContent.swf'],
+          ['back/jokesAndRiddles.swf', update.info.jokes ?? 'archives:News285JokesAndRiddles.swf'],
+          ['back/dividers.swf', update.info.dividersBack ?? 'approximation:dividers_blank.swf'],
+          ['back/navigation.swf', update.info.navigationBack ?? 'archives:News268NavigationBack.swf']
+        ]
+        if (update.info.answers !== undefined) {
+          newspaperComponenets.push(['overlays/riddlesAnswers.swf', update.info.answers]);
+        }
+        if (update.info.extraJokes !== undefined) {
+          newspaperComponenets.push(['overlays/extraJokes.swf', update.info.extraJokes]);
+        }
+        if (update.info.secret !== undefined && update.info.secret !== null) {
+          newspaperComponenets.push(['overlays/secret.swf', update.info.secret]);
+        }
+        if (update.info.iglooWinners !== undefined) {
+          newspaperComponenets.push(['overlays/iglooWinners.swf', update.info.iglooWinners]);
+        }
+        if (update.info.featureMore !== undefined) {
+          newspaperComponenets.push(['overlays/featureMore.swf', update.info.featureMore ?? 'archives:News284FeatureMore.swf']);
+        }
+        if (update.info.supportMore !== undefined) {
+          newspaperComponenets.push(['overlays/supportMore.swf', update.info.supportMore ?? 'archives:News282SupportMore.swf']);
+        }
+        if (update.info.extra !== undefined) {
+          newspaperComponenets.push(['overlays/extra.swf', update.info.extra]);
+        }
+        
+        newspaperComponenets.forEach((pair) => {
+          const [route, file] = pair;
+          this.addRoute(path.join(oldNewsPath, 'content', route), getMediaFilePath(file));
+          this.addRoute(path.join(newNewsPath, 'content', route), getMediaFilePath(file));
+        }) 
+        }
     });
   }
 
@@ -124,6 +203,8 @@ export class GameData {
     this.addIdMap(furnitureDir, 'sprites', FURNITURE_SPRITES);
 
     this.addRouteMap(PRE_CPIP_STATIC_FILES);
+
+    this.addNewspapers();
   }
 
   public addListener(callback: () => void): void {
